requirements.txt
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==6.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
crispy-bootstrap5==2025.4
Django==4.2.20
django-bootstrap-v5==1.0.11
django-crispy-forms==2.4
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-redesign==0.5.1
fontawesomefree==6.6.0
pillow==11.2.1
PyJWT==2.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
